requirements.txt
aiopg==1.2.1
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.5
cffi==1.14.5
channels==3.0.3
channels-postgres==0.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
filelock==3.0.12
django-heroku==0.3.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
install==1.3.4
msgpack==1.0.2
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
virtualenv==20.4.3
zope.interface==5.3.0
djangorestframework~=3.12.4
